+++ src/tables.rs
@@ -1,6 +1,6 @@
 pub const INVALID_VALUE: u8 = 255;
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const STANDARD_ENCODE: &'static [u8; 64] = &[
+#[rustfmt::skip]
+pub const STANDARD_ENCODE: &[u8; 64] = &[
     65, // input 0 (0x0) => 'A' (0x41)
     66, // input 1 (0x1) => 'B' (0x42)
     67, // input 2 (0x2) => 'C' (0x43)
@@ -66,8 +66,8 @@ pub const STANDARD_ENCODE: &'static [u8; 64] = &[
     43, // input 62 (0x3E) => '+' (0x2B)
     47, // input 63 (0x3F) => '/' (0x2F)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const STANDARD_DECODE: &'static [u8; 256] = &[
+#[rustfmt::skip]
+pub const STANDARD_DECODE: &[u8; 256] = &[
     INVALID_VALUE, // input 0 (0x0)
     INVALID_VALUE, // input 1 (0x1)
     INVALID_VALUE, // input 2 (0x2)
@@ -325,8 +325,8 @@ pub const STANDARD_DECODE: &'static [u8; 256] = &[
     INVALID_VALUE, // input 254 (0xFE)
     INVALID_VALUE, // input 255 (0xFF)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const URL_SAFE_ENCODE: &'static [u8; 64] = &[
+#[rustfmt::skip]
+pub const URL_SAFE_ENCODE: &[u8; 64] = &[
     65, // input 0 (0x0) => 'A' (0x41)
     66, // input 1 (0x1) => 'B' (0x42)
     67, // input 2 (0x2) => 'C' (0x43)
@@ -392,8 +392,8 @@ pub const URL_SAFE_ENCODE: &'static [u8; 64] = &[
     45, // input 62 (0x3E) => '-' (0x2D)
     95, // input 63 (0x3F) => '_' (0x5F)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const URL_SAFE_DECODE: &'static [u8; 256] = &[
+#[rustfmt::skip]
+pub const URL_SAFE_DECODE: &[u8; 256] = &[
     INVALID_VALUE, // input 0 (0x0)
     INVALID_VALUE, // input 1 (0x1)
     INVALID_VALUE, // input 2 (0x2)
@@ -651,8 +651,8 @@ pub const URL_SAFE_DECODE: &'static [u8; 256] = &[
     INVALID_VALUE, // input 254 (0xFE)
     INVALID_VALUE, // input 255 (0xFF)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const CRYPT_ENCODE: &'static [u8; 64] = &[
+#[rustfmt::skip]
+pub const CRYPT_ENCODE: &[u8; 64] = &[
     46, // input 0 (0x0) => '.' (0x2E)
     47, // input 1 (0x1) => '/' (0x2F)
     48, // input 2 (0x2) => '0' (0x30)
@@ -718,8 +718,8 @@ pub const CRYPT_ENCODE: &'static [u8; 64] = &[
     121, // input 62 (0x3E) => 'y' (0x79)
     122, // input 63 (0x3F) => 'z' (0x7A)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const CRYPT_DECODE: &'static [u8; 256] = &[
+#[rustfmt::skip]
+pub const CRYPT_DECODE: &[u8; 256] = &[
     INVALID_VALUE, // input 0 (0x0)
     INVALID_VALUE, // input 1 (0x1)
     INVALID_VALUE, // input 2 (0x2)
@@ -977,8 +977,8 @@ pub const CRYPT_DECODE: &'static [u8; 256] = &[
     INVALID_VALUE, // input 254 (0xFE)
     INVALID_VALUE, // input 255 (0xFF)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const IMAP_MUTF7_ENCODE: &'static [u8; 64] = &[
+#[rustfmt::skip]
+pub const IMAP_MUTF7_ENCODE: &[u8; 64] = &[
     65, // input 0 (0x0) => 'A' (0x41)
     66, // input 1 (0x1) => 'B' (0x42)
     67, // input 2 (0x2) => 'C' (0x43)
@@ -1044,8 +1044,8 @@ pub const IMAP_MUTF7_ENCODE: &'static [u8; 64] = &[
     43, // input 62 (0x3E) => '+' (0x2B)
     44, // input 63 (0x3F) => ',' (0x2C)
 ];
-#[cfg_attr(rustfmt, rustfmt_skip)]
-pub const IMAP_MUTF7_DECODE: &'static [u8; 256] = &[
+#[rustfmt::skip]
+pub const IMAP_MUTF7_DECODE: &[u8; 256] = &[
     INVALID_VALUE, // input 0 (0x0)
     INVALID_VALUE, // input 1 (0x1)
     INVALID_VALUE, // input 2 (0x2)